safari-600.7-branch/LayoutTests/java/lc3/CallStatic/boolean-001-expected.txt

LiveConnect 3.0 Call static methods from an instance

PASS dt.staticSetBoolean( true ); dt.PUB_STATIC_BOOLEAN is true
PASS dt.staticSetBoolean( true ); dt.staticGetBoolean() is true
PASS typeof dt.staticGetBoolean() is "boolean"
PASS dt.staticSetBoolean( false ); dt.PUB_STATIC_BOOLEAN is false
PASS dt.staticSetBoolean( false ); dt.staticGetBoolean() is false
PASS typeof dt.staticGetBoolean() is "boolean"
PASS successfullyParsed is true

TEST COMPLETE
 
